+++ R/utils-index-lookup.R
@@ -81,6 +81,8 @@ gen_n <- function(.data) {
 #' @param .data data.frame
 #' @param expr an expression wihch will be interpreted as indices
 #' @param axis 'row' or 'column' expression?
+#'
+#' @noRd
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 loc_expr_to_ids <- function(.data, expr, axis) {
   stopifnot(is.data.frame(.data))